Refactor bunch of code. The font is messed up rn
