+++ src/Components/ImageTwilight.tsx
@@ -1,10 +1,10 @@
 import { Parallax } from "react-parallax";
-import Twilight from "./assets/Twilight/TwilightmasBGIMG.jpeg";
-import Logo from "./assets/Twilight/TwilightMasLogo.png";
-import card1 from "./assets/Twilight/TwilightMasCard1.png";
-import card2 from "./assets/Twilight/TwilightMasCard2.png";
-import card4 from "./assets/Twilight/TwilightMasCard3.png";
-import card3 from "./assets/Twilight/TwilightMasEevee.png";
+import Twilight from "../assets/Twilight/TwilightmasBGIMG.jpeg"
+import Logo from "../assets/Twilight/TwilightMasLogo.png";
+import card1 from "../assets/Twilight/TwilightMasCard1.png";
+import card2 from "../assets/Twilight/TwilightMasCard2.png";
+import card4 from "../assets/Twilight/TwilightMasCard3.png";
+import card3 from "../assets/Twilight/TwilightMasEevee.png";
 
 const ImageTwilight = () => (
   <Parallax className="Pimage" bgImage={Twilight} strength={200}>
